backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
sqlalchemy>=2.0
alembic>=1.13
psycopg[binary]>=3.1
python-dotenv>=1.0
PyJWT>=2.8
pytest>=8.0
